#about redesigned
Adjusted the styling of the .hidden-content class within the #about section to provide a cleaner and more user-friendly design. The background color now features a dark overlay with 75% opacity, and the content is positioned at the bottom with a slight padding. The class is centered within the frame, allowing for improved readability. Additionally, added the ability for vertical scrolling when the content exceeds the viewport height.